Shutting down blog. Thanks for using it :)
This project came to a dead end. It's time to roll the dices again, and start something new. I wasn't able to work on it as long as I've planned, but maybe this is for the best.

Thanks for using blog, this code will be always availible for small blogs and educationnal purposes.